gjk collision check
MJ_D3D11_ConvexHull 을 기반으로 생성된 convex 충돌 체를 사용하여 GJK 충돌 감지 기능이 담긴 MJ_D3D11_GJK 개발
지지함수(Support func)은 MJ_D3D11_Convexhull의 매서드로 개발
현재는 충돌 시 운동 정지 , 이후 EPA 기반의 충돌 대응 개발 예정

+++ d3d11_setup/MJ_D3D11_BasicCam.cpp
@@ -50,7 +50,7 @@ void BasicCam::SetAtV4f(XMVECTOR at)
 
    
 
-void BasicCam::MoveFrontBack(float speed)
+void BasicCam::MoveFrontBack(double speed)
 {
 	XMVECTOR dir = XMVectorSubtract(this->Element.at ,this->Element.pos);// 여기서 w는 0이됨
 
@@ -68,7 +68,7 @@ void BasicCam::MoveFrontBack(float speed)
 }
 
 
-void BasicCam::MoveLeftRight(float speed)
+void BasicCam::MoveLeftRight(double speed)
 {
 	XMVECTOR dir = XMVectorSubtract(this->Element.at, this->Element.pos);// 여기서 w는 0이됨
 
@@ -94,7 +94,7 @@ void BasicCam::MoveLeftRight(float speed)
 
 
 
-void BasicCam::RotToAt(FXMVECTOR axis, float angle)
+void BasicCam::RotToAt(FXMVECTOR axis, double angle)
 {
 	FXMVECTOR quatRotVector = XMQuaternionRotationAxis(axis, angle);
 	XMMATRIX quatRotMat = XMMatrixRotationQuaternion(quatRotVector);
@@ -119,11 +119,11 @@ void BasicCam::CamTranform()
 }
 
 
-void BasicCam::TracballRoate(DirectX::XMINT2 ScreenStart, DirectX::XMINT2 ScreenCur, int ScreenWidth, int ScreenHeight, float RotateRatio)
+void BasicCam::TracballRoate(DirectX::XMINT2 ScreenStart, DirectX::XMINT2 ScreenCur, int ScreenWidth, int ScreenHeight, double RotateRatio)
 {
 	//인풋 정규화 과정
-	XMVECTOR mouseStartV2f = { ((float)ScreenStart.x / ScreenWidth) * 2.F - 1.F ,-1*((float)ScreenStart.y / ScreenHeight) * 2.F + 1.F , 0.F , 0.F };
-	XMVECTOR mouseCurV2f = { ((float)ScreenCur.x / ScreenWidth) * 2.F - 1.F ,-1*((float)ScreenCur.y / ScreenHeight) * 2.F + 1.F , 0.F ,0.F };
+	XMVECTOR mouseStartV2f = { ((double)ScreenStart.x / ScreenWidth) * 2.F - 1.F ,-1*((double)ScreenStart.y / ScreenHeight) * 2.F + 1.F , 0.F , 0.F };
+	XMVECTOR mouseCurV2f = { ((double)ScreenCur.x / ScreenWidth) * 2.F - 1.F ,-1*((double)ScreenCur.y / ScreenHeight) * 2.F + 1.F , 0.F ,0.F };
 
 
 	XMVECTOR moveVector = XMVectorSubtract(mouseCurV2f , mouseStartV2f);
@@ -146,7 +146,7 @@ void BasicCam::TracballRoate(DirectX::XMINT2 ScreenStart, DirectX::XMINT2 Screen
 }
 
 
-void BasicCam::TracballRoateNormVector(XMVECTOR MoveVector, float RotateRatio)
+void BasicCam::TracballRoateNormVector(XMVECTOR MoveVector, double RotateRatio)
 {
 	if (XMVector4Equal(MoveVector, { 0,0,0,0.F })) return;
